notifdjango/requirements.txt
aiofiles==0.7.0
aiohttp==3.7.4.post0
aioredis==1.3.1
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.11.1
Automat==20.2.0
backcall==0.2.0
backports.entry-points-selectable==1.1.0
backports.zoneinfo==0.2.1
bpytop==1.0.67
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
chardet==4.0.0
constantly==15.1.0
cryptography==36.0.1
daphne==3.0.2
decorator==5.1.0
distlib==0.3.3
dj-rest-auth==2.2.3
Django==4.0.1
django-cors-headers==3.11.0
django-crispy-forms==1.13.0
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
filelock==3.3.2
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.11.2
incremental==21.3.0
jedi==0.18.0
Markdown==3.3.6
matplotlib-inline==0.1.3
msgpack==1.0.3
multidict==5.2.0
parso==0.8.2
pickleshare==0.7.5
platformdirs==2.4.0
prompt-toolkit==3.0.20
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.10.0
PyJWT==2.3.0
pyOpenSSL==21.0.0
pytz==2021.3
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
traitlets==5.1.0
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.10.0.2
virtualenv==20.10.0
wcwidth==0.2.5
yarl==1.7.0
zipp==3.7.0
zope.interface==5.4.0
